refactor: Update Kafka broker names and ports in Docker Compose
The code changes in the docker-compose.yaml file update the names and ports of the Kafka brokers. The previous names "kafka-broker-1" and "kafka-broker-2" are changed to "kafka1" and "kafka2" respectively. The ports are also updated from "9092" to "8097" for "kafka1" and from "9093" to "8098" for "kafka2". This change aligns the broker names and ports with the configuration in the Kafka environment variables.

Co-authored-by: Giada De Martino <giada-de-martino@etu.umontpellier.fr>
Co-authored-by: Isalyne <isalyne.llinares-rames@etu.umontpellier.fr>