chore: Update Kafka dependencies and configuration
This commit updates the Kafka dependencies in the Cargo.toml file to include the latest versions of `apache-avro` and `schema_registry_converter`. It also modifies the Kafka initialization and subscription parameters in the `Kafka` struct to improve the configuration. These changes ensure compatibility with the latest Kafka libraries and enhance the performance and functionality of the messaging module.

Co-authored-by: Giada De Martino <giada-de-martino@etu.umontpellier.fr>
Co-authored-by: Isalyne <isalyne.llinares-rames@etu.umontpellier.fr>